# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  github_com_langowen_bodybalance-backend_internal_http-server_admin_admResponse.ErrorResponse:
    properties:
      details:
        description: |-
          Детали ошибки (опционально)
          example: Login field is required
        type: string
      error:
        description: |-
          Текст ошибки
          example: Invalid request
        type: string
    type: object
  github_com_langowen_bodybalance-backend_internal_http-server_admin_admResponse.FileInfo:
    properties:
      mod_time:
        description: |-
          Время последнего изменения
          example: 2023-01-01T12:00:00Z
        type: string
      name:
        description: |-
          Имя файла
          example: video.mp4
        type: string
      size:
        description: |-
          Размер файла в байтах
          example: 1024000
        type: integer
    type: object
...
host: https://api.7375.org
info:
  contact:
    email: info@7375.org
    name: Sergei
  description: API для управления видео-контентом BodyBalance.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: BodyBalance API
  version: "1.0"
paths:
...
  /admin/files/img:
    get:
      description: Возвращает список всех изображений на сервере
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_langowen_bodybalance-backend_internal_http-server_admin_admResponse.FileInfo'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_langowen_bodybalance-backend_internal_http-server_admin_admResponse.ErrorResponse'
      security:
      - AdminAuth: []
      summary: Получить список изображений
      tags:
      - Admin Files
    post:
      consumes:
      - multipart/form-data
      description: Загружает изображение на сервер (макс. 20MB)
      parameters:
      - description: Изображение для загрузки
        in: formData
        name: image
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: Изображение успешно загружено
          schema:
            properties:
              message:
                type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_langowen_bodybalance-backend_internal_http-server_admin_admResponse.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_langowen_bodybalance-backend_internal_http-server_admin_admResponse.ErrorResponse'
      security:
      - AdminAuth: []
      summary: Загрузить изображение
      tags:
      - Admin Files
  /admin/files/video:
    get:
      description: Возвращает список всех видеофайлов на сервере
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_langowen_bodybalance-backend_internal_http-server_admin_admResponse.FileInfo'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_langowen_bodybalance-backend_internal_http-server_admin_admResponse.ErrorResponse'
      security:
      - AdminAuth: []
      summary: Получить список видеофайлов
      tags:
      - Admin Files
    post:
      consumes:
      - multipart/form-data
      description: Загружает видеофайл на сервер (макс. 500MB)
      parameters:
      - description: Видеофайл для загрузки
        in: formData
        name: video
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: Файл успешно загружен
          schema:
            properties:
              message:
                type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_langowen_bodybalance-backend_internal_http-server_admin_admResponse.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_langowen_bodybalance-backend_internal_http-server_admin_admResponse.ErrorResponse'
      security:
      - AdminAuth: []
      summary: Загрузить видеофайл
      tags:
      - Admin Files
...
swagger: "2.0"